requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.21.1
langchain==0.0.354
langchain-community==0.0.8
dashscope==1.24.6
openai==1.3.5
python-dotenv==1.0.0
pydantic==1.10.13
PyPDF2==3.0.1
python-docx==0.8.11
python-multipart==0.0.6
# 向量数据库相关依赖
qdrant-client==1.9.1
sentence-transformers==2.2.2
# 用户认证相关依赖
sqlalchemy==2.0.23
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
# MySQL数据库依赖
pymysql==1.0.2
# 语音识别和合成依赖
SpeechRecognition==3.10.1
pyttsx3==2.90
PyAudio==0.2.13